Separated MixedBody from ForcePoint.cs
